requirements.txt
asgiref==3.7.2
asttokens==2.4.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.7
Django==4.2.6
executing==2.0.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==3.11.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
setuptools==73.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.13.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.2.79
ultralytics-thop==2.0.5
urllib3==2.2.2
virtualenv==20.24.6
wcwidth==0.2.13
Werkzeug==3.0.3
